Fixes bugs in ground truth files (and in the type mismatch finding code)

+++ src/main_analyze_results.py
@@ -93,11 +93,7 @@ def compare_json_files(expected, out):
             #     for _type in fact_expected.get("type", []):
             #         if _type in fact_out.get("type", []):
             #             partial_matches += 1
-            elif all(
-                x in fact_out and fact_out[x] == fact_expected[x]
-                for x in fact_expected.keys()
-                if x not in ["type", "col_offset"]
-            ):
+            elif utils.is_same_element(fact_out, fact_expected):
                 entry_match = True
                 break
             else:
@@ -131,7 +127,8 @@ def format_missing_matches(all_missing_matches):
         "File",
         "Line Number",
         "Function",
-        "Parameter/Variable",
+        "Parameter",
+        "Variable",
         "Actual Type",
         "Predicted Type",
     ]
@@ -193,7 +190,7 @@ def format_missing_matches(all_missing_matches):
     ]
     formatted_output = (
         "\nMissing matches:\n{}\nTotal error entries: {}\nErrors in function return"
-        " type: {}\nErrors in param: {}\n/Errors in variables:{}\nMissing entries:"
+        " type: {}\nErrors in param: {}\nErrors in variables: {}\nMissing entries:"
         " {}\nMismatch entries: {}".format(
             missing_matches_table,
             total_missing_entries,
@@ -1048,14 +1045,16 @@ def run_results_analyzer(results_dir, benchmark_dir="micro-benchmark"):
     )
 
     for tool in list(tools_results.keys()):
-        shutil.move(
-            f"{tool}_mismatches_reasons.csv",
-            f"{str(results_dir)}/mismatches/{tool}_mismatches_reasons.csv",
-        )
-        shutil.move(
-            f"{tool}_not_found_reasons.csv",
-            f"{str(results_dir)}/missing/{tool}_not_found_reasons.csv",
-        )
+        if Path(f"{tool}_mismatches_reasons.csv").exists():
+            shutil.move(
+                f"{tool}_mismatches_reasons.csv",
+                f"{str(results_dir)}/mismatches/{tool}_mismatches_reasons.csv",
+            )
+        if Path(f"{tool}_not_found_reasons.csv").exists():
+            shutil.move(
+                f"{tool}_not_found_reasons.csv",
+                f"{str(results_dir)}/missing/{tool}_not_found_reasons.csv",
+            )
 
         if tool in utils.ML_TOOLS:
             shutil.move(
